refactoring : 코드 리뷰 적용
